requirements.txt
aiomysql==0.0.21
aiosqlite==0.16.1
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
fastapi==0.65.2
h11==0.12.0
httptools==0.2.0
idna==2.10
iso8601==0.1.14
itsdangerous==2.0.1
pydantic==1.8.2
PyMySQL==0.9.3
pypika-tortoise==0.1.1
python-dotenv==0.18.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
starlette==0.14.2
tortoise-orm==0.17.4
typing-extensions==3.10.0.0
ujson==4.0.2
urllib3==1.26.5
uvicorn==0.14.0
watchgod==0.7
websockets==9.1
